+++ crates/swc_ecma_parser/src/lib.rs
@@ -150,12 +150,12 @@ mod parser;
 pub enum Syntax {
     /// Standard
     #[serde(rename = "ecmascript")]
-    Es(EsConfig),
+    Es(EsSyntax),
     /// This variant requires the cargo feature `typescript` to be enabled.
     #[cfg(feature = "typescript")]
     #[cfg_attr(docsrs, doc(cfg(feature = "typescript")))]
     #[serde(rename = "typescript")]
-    Typescript(TsConfig),
+    Typescript(TsSyntax),
 }
 
 impl Default for Syntax {
@@ -167,7 +167,7 @@ impl Default for Syntax {
 impl Syntax {
     fn auto_accessors(self) -> bool {
         match self {
-            Syntax::Es(EsConfig {
+            Syntax::Es(EsSyntax {
                 auto_accessors: true,
                 ..
             }) => true,
@@ -179,7 +179,7 @@ impl Syntax {
 
     pub fn import_attributes(self) -> bool {
         match self {
-            Syntax::Es(EsConfig {
+            Syntax::Es(EsSyntax {
                 import_attributes, ..
             }) => import_attributes,
             #[cfg(feature = "typescript")]
@@ -190,24 +190,24 @@ impl Syntax {
     /// Should we parse jsx?
     pub fn jsx(self) -> bool {
         match self {
-            Syntax::Es(EsConfig { jsx: true, .. }) => true,
+            Syntax::Es(EsSyntax { jsx: true, .. }) => true,
             #[cfg(feature = "typescript")]
-            Syntax::Typescript(TsConfig { tsx: true, .. }) => true,
+            Syntax::Typescript(TsSyntax { tsx: true, .. }) => true,
             _ => false,
         }
     }
 
     pub fn fn_bind(self) -> bool {
-        matches!(self, Syntax::Es(EsConfig { fn_bind: true, .. }))
+        matches!(self, Syntax::Es(EsSyntax { fn_bind: true, .. }))
     }
 
     pub fn decorators(self) -> bool {
         match self {
-            Syntax::Es(EsConfig {
+            Syntax::Es(EsSyntax {
                 decorators: true, ..
             }) => true,
             #[cfg(feature = "typescript")]
-            Syntax::Typescript(TsConfig {
+            Syntax::Typescript(TsSyntax {
                 decorators: true, ..
             }) => true,
             _ => false,
@@ -216,7 +216,7 @@ impl Syntax {
 
     pub fn decorators_before_export(self) -> bool {
         match self {
-            Syntax::Es(EsConfig {
+            Syntax::Es(EsSyntax {
                 decorators_before_export: true,
                 ..
             }) => true,
@@ -241,7 +241,7 @@ impl Syntax {
     pub fn export_default_from(self) -> bool {
         matches!(
             self,
-            Syntax::Es(EsConfig {
+            Syntax::Es(EsSyntax {
                 export_default_from: true,
                 ..
             })
@@ -258,7 +258,7 @@ impl Syntax {
 
     pub(crate) fn allow_super_outside_method(self) -> bool {
         match self {
-            Syntax::Es(EsConfig {
+            Syntax::Es(EsSyntax {
                 allow_super_outside_method,
                 ..
             }) => allow_super_outside_method,
@@ -269,7 +269,7 @@ impl Syntax {
 
     pub(crate) fn allow_return_outside_function(self) -> bool {
         match self {
-            Syntax::Es(EsConfig {
+            Syntax::Es(EsSyntax {
                 allow_return_outside_function,
                 ..
             }) => allow_return_outside_function,
@@ -296,7 +296,7 @@ impl Syntax {
 
     pub fn explicit_resource_management(&self) -> bool {
         match self {
-            Syntax::Es(EsConfig {
+            Syntax::Es(EsSyntax {
                 explicit_resource_management: using_decl,
                 ..
             }) => *using_decl,
@@ -308,7 +308,7 @@ impl Syntax {
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Default, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
-pub struct TsConfig {
+pub struct TsSyntax {
     #[serde(default)]
     pub tsx: bool,
 
@@ -331,9 +331,12 @@ pub struct TsConfig {
     pub disallow_ambiguous_jsx_like: bool,
 }
 
+#[deprecated(note = "Use 'TsSyntax' instead")]
+pub type TsConfig = TsSyntax;
+
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Default, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
-pub struct EsConfig {
+pub struct EsSyntax {
     #[serde(default)]
     pub jsx: bool,
 
@@ -372,6 +375,9 @@ pub struct EsConfig {
     #[serde(default)]
     pub explicit_resource_management: bool,
 }
+
+#[deprecated(note = "Use 'EsSyntax' instead")]
+pub type EsConfig = EsSyntax;
 
 /// Syntactic context.
 #[derive(Debug, Clone, Copy, Default)]