+++ crates/swc_ecma_transforms_proposal/tests/export.rs
@@ -1,12 +1,12 @@
 use swc_common::chain;
-use swc_ecma_parser::{EsConfig, Syntax};
+use swc_ecma_parser::{EsSyntax, Syntax};
 use swc_ecma_transforms_compat::es2020::export_namespace_from;
 use swc_ecma_transforms_proposal::export_default_from;
 use swc_ecma_transforms_testing::test;
 use swc_ecma_visit::Fold;
 
 fn syntax_default() -> Syntax {
-    Syntax::Es(EsConfig {
+    Syntax::Es(EsSyntax {
         export_default_from: true,
         ..Default::default()
     })