+++ crates/swc_ecma_transforms_react/src/jsx_self/tests.rs
@@ -7,7 +7,7 @@ fn tr() -> impl Fold {
 }
 
 test!(
-    ::swc_ecma_parser::Syntax::Es(::swc_ecma_parser::EsConfig {
+    ::swc_ecma_parser::Syntax::Es(::swc_ecma_parser::EsSyntax {
         jsx: true,
         ..Default::default()
     }),