+++ crates/swc_estree_compat/tests/convert.rs
@@ -16,7 +16,7 @@ use swc_common::{
     errors::{ColorConfig, Handler},
     FileName, FilePathMapping, SourceMap, GLOBALS,
 };
-use swc_ecma_parser::{EsConfig, Syntax};
+use swc_ecma_parser::{EsSyntax, Syntax};
 use swc_estree_compat::babelify::{Babelify, Context};
 use test::{test_main, DynTestFn, ShouldPanic, TestDesc, TestDescAndFn, TestName, TestType};
 use testing::{json::diff_json_value, DebugUsingDisplay};
@@ -85,7 +85,7 @@ fn fixtures() -> Result<(), Error> {
                     let syntax = if is_typescript {
                         Syntax::Typescript(Default::default())
                     } else if is_jsx {
-                        Syntax::Es(EsConfig {
+                        Syntax::Es(EsSyntax {
                             jsx: true,
                             ..Default::default()
                         })